Fixed Return Type

+++ src/GreenSale.DataAccess/Interfaces/BuyerPosts/IBuyerPostStarRepository.cs
@@ -5,6 +5,6 @@ namespace GreenSale.DataAccess.Interfaces.BuyerPosts;
 
 public interface IBuyerPostStarRepository : IRepository<BuyerPostStars,BuyerPostStars>
 {
-    public Task<List<long>> GetAllStarsPostIdCountAsync(long postid);
+    public Task<List<int>> GetAllStarsPostIdCountAsync(long postid);
     public Task<long> GetIdAsync(long userid, long postid);
 }
